removed semicolons from REFERENCE_PATTERN
When parsing text with multiple references in which one of the references begins with a numeral (i.e. "Gal. 6:2; Hebrews 3:6; 2 Cor. 2:9"), the presence of a semicolon in the REFERENCE_PATTERN cause the regex to gobble up the number of the last reference. In other words, using the text noted previously, only to pericopes would be found. Hebrews 3:6 would be found as Hebrews 3:6, 2. The last reference of 2 Cor. 2:9 would not be found.

Perhaps there is a better way to adjust the regex other than deleting the semicolons but I haven't yet figured it out.